feat(order): update order and command repositories to use Integer IDs and add new use cases

+++ src/main/java/development/v/development/application/features/CashReceipt/CreateCashReceipt/CreateCashReceiptUseCase.java
@@ -46,7 +46,7 @@ public class CreateCashReceiptUseCase {
         customerRepository.findById(request.getCliId())
                 .orElseThrow(() -> new NotFoundException(Message.NOT_FOUND + ": Cliente " + request.getCliId()));
 
-        orderRepository.findById(request.getPedId())
+        orderRepository.findById(request.getPedId().intValue())
                 .orElseThrow(() -> new NotFoundException(Message.NOT_FOUND + ": Pedido " + request.getPedId()));
     }
 }
